backend/requirements.txt
fastapi>=0.115.0
python-dotenv>=1.0.0
supabase>=2.0.0
uvicorn[standard]>=0.24.0
